Pull OSFileStore getter boilerplate into AbstractOSFileStore
Move logicalVolume, description, fsType, freeSpace, usableSpace,
totalSpace, freeInodes, and totalInodes fields plus their getters
and updateFrom/updateSpaceAndInodes/updateSpace helpers into the
abstract base class.

Simplifies 7 subclasses by removing ~57 lines of duplicated
fields+getters from each.

Closes partially: #3274
