+++ oshi-common/src/main/java/oshi/software/common/os/linux/LinuxOSFileStore.java
@@ -15,15 +15,6 @@ import oshi.software.os.OSFileStore;
 public class LinuxOSFileStore extends AbstractOSFileStore {
 
     private final LinuxFileSystem fs;
-    private String logicalVolume;
-    private String description;
-    private String fsType;
-
-    private long freeSpace;
-    private long usableSpace;
-    private long totalSpace;
-    private long freeInodes;
-    private long totalInodes;
 
     /**
      * Creates a LinuxOSFileStore.
@@ -48,56 +39,9 @@ public class LinuxOSFileStore extends AbstractOSFileStore {
     public LinuxOSFileStore(String name, String volume, String label, String mount, String options, String uuid,
             boolean local, String logicalVolume, String description, String fsType, long freeSpace, long usableSpace,
             long totalSpace, long freeInodes, long totalInodes, LinuxFileSystem fs) {
-        super(name, volume, label, mount, options, uuid, local);
+        super(name, volume, label, mount, options, uuid, local, logicalVolume, description, fsType, freeSpace,
+                usableSpace, totalSpace, freeInodes, totalInodes);
         this.fs = fs;
-        this.logicalVolume = logicalVolume;
-        this.description = description;
-        this.fsType = fsType;
-        this.freeSpace = freeSpace;
-        this.usableSpace = usableSpace;
-        this.totalSpace = totalSpace;
-        this.freeInodes = freeInodes;
-        this.totalInodes = totalInodes;
-    }
-
-    @Override
-    public String getLogicalVolume() {
-        return this.logicalVolume;
-    }
-
-    @Override
-    public String getDescription() {
-        return this.description;
-    }
-
-    @Override
-    public String getType() {
-        return this.fsType;
-    }
-
-    @Override
-    public long getFreeSpace() {
-        return this.freeSpace;
-    }
-
-    @Override
-    public long getUsableSpace() {
-        return this.usableSpace;
-    }
-
-    @Override
-    public long getTotalSpace() {
-        return this.totalSpace;
-    }
-
-    @Override
-    public long getFreeInodes() {
-        return this.freeInodes;
-    }
-
-    @Override
-    public long getTotalInodes() {
-        return this.totalInodes;
     }
 
     @Override
@@ -126,25 +70,5 @@ public class LinuxOSFileStore extends AbstractOSFileStore {
             }
         }
         return false;
-    }
-
-    private void updateSpaceAndInodes(long freeSpace, long usableSpace, long totalSpace, long freeInodes,
-            long totalInodes) {
-        this.freeSpace = freeSpace;
-        this.usableSpace = usableSpace;
-        this.totalSpace = totalSpace;
-        this.freeInodes = freeInodes;
-        this.totalInodes = totalInodes;
-    }
-
-    private void updateFrom(OSFileStore fileStore) {
-        this.logicalVolume = fileStore.getLogicalVolume();
-        this.description = fileStore.getDescription();
-        this.fsType = fileStore.getType();
-        this.freeSpace = fileStore.getFreeSpace();
-        this.usableSpace = fileStore.getUsableSpace();
-        this.totalSpace = fileStore.getTotalSpace();
-        this.freeInodes = fileStore.getFreeInodes();
-        this.totalInodes = fileStore.getTotalInodes();
     }
 }
